Move all mixins to app related mixins files
