+++ backend/posts_comments/serializers.py
@@ -1,22 +1,9 @@
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
-from events.serializers import ValidateUserInContextMixin
+from events.mixins import ValidateUserInContextMixin
 from posts_comments import models
-
-
-class VoteMixin:
-    def get_vote_count(self, obj):
-        return obj.votes.all().count()
-
-    def get_has_voted(self, obj):
-        request = self.context.get("request", None)
-
-        if not hasattr(request, "user") or not request.user.is_authenticated:
-            return False
-
-        votes = obj.votes.all()
-        return votes.filter(user=request.user).exists()
+from posts_comments.mixins import VoteMixin
 
 
 class PostSerializer(VoteMixin,
